change tracking and .dump support
I believe that I now more or less properly track changes that gdrive
sends down.

Also added mnt/.dump support which is a useful debugging tool to get an
idea of what all the metadata the fs thinks it has.

Need to understand caching better.  When/how does the kernel cache
attributes and/or content?  How can I take advantage of that and work
with it?