file content management work
. Now instead of reading the entire file into memory and returning it
via ReadDirAll, we: fetch it into a temporary file, then return chunks
of it on demand, cleaning it up when our file is released.
. We also tell the kernel it is ok to cache file contents and we
invalidate them when something changes.

Currently cache invalidation is super aggressive and we invalidate
things if there is any change at all, even, for example, a change in the
last view time for a file.